+++ mrserver/handler_adapter.go
@@ -1,0 +1,29 @@
+package mrserver
+
+import (
+    "net/http"
+
+    "github.com/mondegor/go-webcore/mrcore"
+    "github.com/mondegor/go-webcore/mrctx"
+)
+
+func HandlerAdapter(validator mrcore.Validator) mrcore.HttpHandlerAdapterFunc {
+    return func(next mrcore.HttpHandlerFunc) http.HandlerFunc {
+        return func(w http.ResponseWriter, r *http.Request) {
+            logger := mrctx.Logger(r.Context())
+            logger.Debug("Exec HandlerAdapter")
+
+            c := clientContext{
+                request: r,
+                responseWriter: w,
+                validator: validator,
+            }
+
+            err := next(&c)
+
+            if err != nil {
+                c.sendErrorResponse(err)
+            }
+        }
+    }
+}