+++ mrtool/app_helper.go
@@ -1,4 +1,4 @@
-package mrcore
+package mrtool
 
 import (
     "context"
@@ -7,15 +7,17 @@ import (
     "os"
     "os/signal"
     "syscall"
+
+    "github.com/mondegor/go-webcore/mrcore"
 )
 
 type (
 	AppHelper struct {
-        logger Logger
+        logger mrcore.Logger
     }
 )
 
-func NewAppHelper(logger Logger) *AppHelper {
+func NewAppHelper(logger mrcore.Logger) *AppHelper {
     return &AppHelper{
         logger: logger,
     }
@@ -32,7 +34,7 @@ func (h *AppHelper) Close(c io.Closer) {
     err := c.Close()
 
     if err != nil {
-        h.logger.Err(FactoryErrInternalFailedToClose.Caller(1).Wrap(err, fmt.Sprintf("%v", c)))
+        h.logger.Err(mrcore.FactoryErrInternalFailedToClose.Caller(1).Wrap(err, fmt.Sprintf("%v", c)))
     }
 }